Minimal benchmarking suite
SharpBench provides a way to profile static Actions (functions which
take and return no argument) in an assembly simply by marking them
with the Bench attribute.

BenchRunner visits the static methods and executes them multiple times
to try and get an estimate of the time spent in the function and
decompose it into computation and garbage collection.